+++ MTP/interface/locations/LocationsVC.swift
@@ -37,6 +37,7 @@ class LocationsVC: UIViewController {
     private var matches: [Mappable] = []
 
     private var injectMappable: Mappable?
+    private var injectCenter: CLLocationCoordinate2D?
 
     /// :nodoc:
     deinit {
@@ -54,7 +55,6 @@ class LocationsVC: UIViewController {
     }
 
     /// Prepare for reveal
-    ///
     /// - Parameter animated: Whether animating
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
@@ -64,7 +64,6 @@ class LocationsVC: UIViewController {
     }
 
     /// Actions to take after reveal
-    ///
     /// - Parameter animated: Whether animating
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
@@ -76,20 +75,31 @@ class LocationsVC: UIViewController {
     }
 
     /// Instrument and inject navigation
-    ///
     /// - Parameters:
     ///   - segue: Navigation action
     ///   - sender: Action originator
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
         if let nearby = Segues.showNearby(segue: segue)?
                               .destination {
-            nearby.inject(model: (data.visibles,
-                                  mtpMapView.centerCoordinate))
+            let center = injectCenter ?? mtpMapView.centerCoordinate
+            injectCenter = nil
+            nearby.inject(model: (data.visibles, center))
         } else if let show = Segues.showLocation(segue: segue)?
                                    .destination,
                   let inject = injectMappable {
             show.inject(model: inject)
             injectMappable = nil
+        } else if let photo = Segues.addPhoto(segue: segue)?
+                                    .destination {
+            photo.inject(model: (mappable: injectMappable,
+                                 delegate: self))
+            injectMappable = nil
+        } else if let post = Segues.addPost(segue: segue)?
+                                   .destination {
+            if let mappable = injectMappable {
+                post.inject(model: mappable)
+                injectMappable = nil
+            }
         }
     }
 
@@ -99,7 +109,6 @@ class LocationsVC: UIViewController {
     }
 
     /// Reveal user
-    ///
     /// - Parameter user: User to display
     func reveal(user: User?) {
         guard let name = user?.locationName, !name.isEmpty else { return }
@@ -117,15 +126,29 @@ class LocationsVC: UIViewController {
 
 extension LocationsVC: Mapper {
 
-    /// Close
-    ///
+    /// Show Add Photo screen
+    /// - Parameter mappable: Place
+    func add(photo mappable: Mappable) {
+        injectMappable = mappable
+        performSegue(withIdentifier: Segues.addPhoto,
+                     sender: self)
+    }
+
+    /// Show Add Post screen
+    /// - Parameter mappable: Place
+    func add(post mappable: Mappable) {
+        injectMappable = mappable
+        performSegue(withIdentifier: Segues.addPost,
+                     sender: self)
+    }
+
+    /// Close callout
     /// - Parameter mappable: Place
     func close(mappable: Mappable) {
         mtpMapView.close(mappable: mappable)
     }
 
-    /// Notify
-    ///
+    /// Notify of visit
     /// - Parameters:
     ///   - mappable: Place
     ///   - triggered: Date
@@ -134,8 +157,7 @@ extension LocationsVC: Mapper {
                     triggered: triggered) { _ in }
     }
 
-    /// Reveal
-    ///
+    /// Reveal on map
     /// - Parameters:
     ///   - mappable: Place
     ///   - callout: Show callout
@@ -144,18 +166,24 @@ extension LocationsVC: Mapper {
         mtpMapView.zoom(to: mappable, callout: callout)
     }
 
-    /// Show
-    ///
+    /// Show Show More screen
     /// - Parameter mappable: Place
-    func show(mappable: Mappable) {
+    func show(more mappable: Mappable) {
         injectMappable = mappable
         close(mappable: mappable)
         performSegue(withIdentifier: Segues.showLocation,
                      sender: self)
     }
 
+    /// Show Nearby screen
+    /// - Parameter mappable: Place
+    func show(nearby mappable: Mappable) {
+        injectCenter = mappable.coordinate
+        performSegue(withIdentifier: Segues.showNearby,
+                     sender: self)
+    }
+
     /// Update
-    ///
     /// - Parameter mappable: Place
     func update(mappable: Mappable) {
         mtpMapView.update(mappable: mappable)
@@ -172,14 +200,12 @@ extension LocationsVC: LocationTracker {
     }
 
     /// Authorization changed
-    ///
     /// - Parameter changed: New status
     func authorization(changed: CLAuthorizationStatus) {
         updateTracking()
     }
 
     /// Location changed
-    ///
     /// - Parameter changed: New location
     func location(changed: CLLocation) { }
 }
@@ -254,7 +280,6 @@ extension LocationsVC: Exposing {
 extension LocationsVC: UISearchBarDelegate {
 
     /// Changed search text notification
-    ///
     /// - Parameters:
     ///   - searchBar: Searcher
     ///   - searchText: Contents
@@ -275,21 +300,18 @@ extension LocationsVC: UISearchBarDelegate {
     }
 
     /// Begin search editing
-    ///
     /// - Parameter searchBar: Searcher
     func searchBarTextDidBeginEditing(_ searchBar: UISearchBar) {
         searchBar.setShowsCancelButton(true, animated: true)
     }
 
     /// Handle search button click
-    ///
     /// - Parameter searchBar: Searcher
     func searchBarSearchButtonClicked(_ searchBar: UISearchBar) {
         searchBarCancelButtonClicked(searchBar)
     }
 
     /// Handle cancel button click
-    ///
     /// - Parameter searchBar: Searcher
     func searchBarCancelButtonClicked(_ searchBar: UISearchBar) {
         searchBar.text = ""
@@ -298,7 +320,6 @@ extension LocationsVC: UISearchBarDelegate {
     }
 
     /// Search ended notification
-    ///
     /// - Parameter searchBar: Searcher
     func searchBarTextDidEndEditing(_ searchBar: UISearchBar) {
         searchBar.showsCancelButton = false
@@ -310,7 +331,6 @@ extension LocationsVC: UISearchBarDelegate {
 extension LocationsVC: MKMapViewDelegate {
 
     /// Update user location
-    ///
     /// - Parameters:
     ///   - mapView: Map view
     ///   - userLocation: Location
@@ -320,7 +340,6 @@ extension LocationsVC: MKMapViewDelegate {
     }
 
     /// Produce annotation view
-    ///
     /// - Parameters:
     ///   - mapView: Map view
     ///   - annotation: Annotation
@@ -341,7 +360,6 @@ extension LocationsVC: MKMapViewDelegate {
     }
 
     /// Handle annoation selection
-    ///
     /// - Parameters:
     ///   - mapView: Map view
     ///   - view: Annotation view
@@ -359,7 +377,6 @@ extension LocationsVC: MKMapViewDelegate {
     }
 
     /// Provide overlay renderer
-    ///
     /// - Parameters:
     ///   - mapView: Map view
     ///   - overlay: Overlay
@@ -373,6 +390,14 @@ extension LocationsVC: MKMapViewDelegate {
             return MKOverlayRenderer(overlay: overlay)
         }
     }
+}
+
+// MARK: - MKUserTrackingButton
+
+extension LocationsVC: AddPhotoDelegate {
+
+    /// Enable Location selection
+    var isLocatable: Bool { return true }
 }
 
 // MARK: - MKUserTrackingButton